requirements.txt
Django==1.9.7
django-debug-toolbar==1.4
django-rest-swagger==2.0.1
djangorestframework==3.4.0
mysqlclient==1.3.7
mysqlclient==3.1.23
